FHAC-623 PR updated with review comments
